Refactor compute_satellite_visibility to use more succinct code.
Create new RhoVector objects for each overpass, or take a slice of an existing one.

Pass in the object ECEF vector and location information to RhoVector to get the SEZ vector.

Create the SatelliteVisibility named tuple for output of the function

Change the Location schema to not use aliases, rely on the abbreviated names for attributes

Use a basic LRU cache on the site_ECEF function since it gets called multiple times for the same inputs